Reorganize constraints and transforms modules (#2052)
