+++ examples/dmm/dmm.py
@@ -24,7 +24,8 @@ import polyphonic_data_loader as poly
 import pyro
 import pyro.distributions as dist
 import pyro.poutine as poutine
-from pyro.distributions import InverseAutoregressiveFlow, TransformedDistribution
+from pyro.distributions import TransformedDistribution
+from pyro.distributions.transforms import InverseAutoregressiveFlow
 from pyro.infer import SVI, JitTrace_ELBO, Trace_ELBO
 from pyro.nn import AutoRegressiveNN
 from pyro.optim import ClippedAdam
