+++ pyro/distributions/__init__.py
@@ -1,5 +1,7 @@
 import pyro.distributions.torch_patch  # noqa F403
 from pyro.distributions.avf_mvn import AVFMultivariateNormal
+from pyro.distributions.conditional import (ConditionalDistribution, ConditionalTransform,
+                                            ConditionalTransformedDistribution, ConditionalTransformModule)
 from pyro.distributions.conjugate import BetaBinomial, DirichletMultinomial, GammaPoisson
 from pyro.distributions.delta import Delta
 from pyro.distributions.diag_normal_mixture import MixtureOfDiagNormals
@@ -9,6 +11,7 @@ from pyro.distributions.empirical import Empirical
 from pyro.distributions.gaussian_scale_mixture import GaussianScaleMixture
 from pyro.distributions.hmm import DiscreteHMM, GaussianHMM, GaussianMRF
 from pyro.distributions.inverse_gamma import InverseGamma
+from pyro.distributions.lkj import LKJCorrCholesky
 from pyro.distributions.mixture import MaskedMixture
 from pyro.distributions.omt_mvn import OMTMultivariateNormal
 from pyro.distributions.rejector import Rejector
@@ -19,25 +22,21 @@ from pyro.distributions.torch import *  # noqa F403
 from pyro.distributions.torch import __all__ as torch_dists
 from pyro.distributions.torch_distribution import TorchDistribution
 from pyro.distributions.torch_transform import TransformModule
-from pyro.distributions.conditional import (ConditionalDistribution, ConditionalTransform, ConditionalTransformModule,
-                                            ConditionalTransformedDistribution)
-from pyro.distributions.util import enable_validation, is_validation_enabled, validation_enabled
 from pyro.distributions.unit import Unit
+from pyro.distributions.util import enable_validation, is_validation_enabled, validation_enabled
 from pyro.distributions.von_mises import VonMises
 from pyro.distributions.von_mises_3d import VonMises3D
 from pyro.distributions.zero_inflated_poisson import ZeroInflatedPoisson
-from pyro.distributions.lkj import (LKJCorrCholesky, CorrLCholeskyTransform, corr_cholesky_constraint)
-from pyro.distributions.transforms import *  # noqa F403
 
-import pyro.distributions.kl  # noqa F403 isort:skip
+from . import constraints, kl, transforms
 
 __all__ = [
     "AVFMultivariateNormal",
     "BetaBinomial",
     "ConditionalDistribution",
     "ConditionalTransform",
-    "ConditionalTransformedDistribution",
     "ConditionalTransformModule",
+    "ConditionalTransformedDistribution",
     "Delta",
     "DirichletMultinomial",
     "DiscreteHMM",
@@ -49,8 +48,6 @@ __all__ = [
     "GaussianScaleMixture",
     "InverseGamma",
     "LKJCorrCholesky",
-    "CorrLCholeskyTransform",
-    "corr_cholesky_constraint",
     "MaskedMixture",
     "MixtureOfDiagNormals",
     "MixtureOfDiagNormalsSharedCovariance",
@@ -65,14 +62,14 @@ __all__ = [
     "VonMises",
     "VonMises3D",
     "ZeroInflatedPoisson",
+    "constraints",
     "enable_validation",
     "is_validation_enabled",
+    "kl",
+    "transforms",
     "validation_enabled",
 ]
 
 # Import all torch distributions from `pyro.distributions.torch_distribution`
 __all__.extend(torch_dists)
 del torch_dists
-
-# For backwards compatibility, import all transforms into distributions module
-__all__.extend(transforms.__all__)  # noqa
